requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.2
idna==2.6
oauthlib==2.0.6
Pillow==5.0.0
psycopg2==2.7.4
PyJWT==1.5.3
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
urllib3==1.22
django-leaflet==0.23.0
django-geojson==2.11.0
django-datetime-widget==0.9.3